+++ condenser/tests/conftest.py
@@ -1,18 +1,37 @@
 from .schema import Base
 from .schema import ModelOne
 from condenser import NumpyQuery
+from condenser.utils import load_spatialite
 from sqlalchemy import create_engine
 from sqlalchemy.orm import scoped_session
 from sqlalchemy.orm import sessionmaker
+from sqlalchemy.event import listen
+from sqlalchemy.sql import select, func
 
-import os
 import pytest
+
+try:
+    from geoalchemy2.types import Geometry  # NOQA
+    import pygeos  # NOQA
+
+    has_geo = True
+except ImportError:
+    has_geo = False
+
+requires_geo = pytest.mark.skipif(not has_geo, reason="requires GeoAlchemy2 and pygeos")
 
 
 @pytest.fixture(scope="session")
 def db_engine(request):
     """yields a SQLAlchemy engine which is suppressed after the test session"""
     engine = create_engine("sqlite://")
+    if has_geo:
+        # https://geoalchemy-2.readthedocs.io/en/latest/spatialite_tutorial.html
+        listen(engine, "connect", load_spatialite)
+        conn = engine.connect()
+        conn.execute(select([func.InitSpatialMetaData()]))
+        conn.close()
+
     session_factory = scoped_session(sessionmaker(bind=engine))
 
     Base.metadata.create_all(engine)
@@ -24,6 +43,8 @@ def db_engine(request):
         col_bool=True,
         col_text="once upon a time",
     )
+    if has_geo:
+        record.col_geom = "POINT (2 3)"
 
     session = session_factory()
     session.add(record)
@@ -37,13 +58,13 @@ def db_engine(request):
 @pytest.fixture(scope="session")
 def db_session_factory(db_engine):
     """returns a SQLAlchemy scoped session factory"""
-    return scoped_session(sessionmaker(bind=db_engine))
+    return scoped_session(sessionmaker(bind=db_engine, query_cls=NumpyQuery))
 
 
-@pytest.fixture(scope="session")
+@pytest.fixture(scope="function")
 def db_session(db_session_factory):
     """yields a SQLAlchemy connection which is rollbacked after the test"""
-    session = db_session_factory(query_cls=NumpyQuery)
+    session = db_session_factory()
 
     yield session