+++ condenser/tests/schema.py
@@ -6,7 +6,6 @@ from sqlalchemy import String
 from sqlalchemy import Text
 from sqlalchemy.ext.declarative import declarative_base
 
-
 Base = declarative_base()
 
 
@@ -18,3 +17,10 @@ class ModelOne(Base):
     col_str = Column(String(32))
     col_text = Column(Text)
     col_float = Column(Float)
+
+    try:
+        from geoalchemy2.types import Geometry
+
+        col_geom = Column(Geometry(geometry_type="POINT", management=True))
+    except ImportError:
+        pass